Improve impl

+++ src/sentry/tasks/assemble.py
@@ -335,14 +335,17 @@ def _bind_or_create_artifact_bundle(
         # We store a reference to the previous file to which the bundle was pointing to.
         existing_file = existing_artifact_bundle.file
 
-        # In case there is an ArtifactBundle with a specific bundle_id, we want to change its underlying File model
-        # with its corresponding artifact count.
-        existing_artifact_bundle.update(
-            date_added=date_added, file=archive_file, artifact_count=artifact_count
-        )
+        # Only if the file objects are different we want to update the database, otherwise we will end up deleting
+        # a newly bound file.
+        if existing_file != archive_file:
+            # In case there is an ArtifactBundle with a specific bundle_id, we want to change its underlying File model
+            # with its corresponding artifact count.
+            existing_artifact_bundle.update(
+                date_added=date_added, file=archive_file, artifact_count=artifact_count
+            )
 
-        # We now delete that file, in order to avoid orphan files in the database.
-        existing_file.delete()
+            # We now delete that file, in order to avoid orphan files in the database.
+            existing_file.delete()
 
         return existing_artifact_bundle, False
